+++ Common/src/main/java/fuzs/bagofholding/api/config/ContainerItemTooltipConfig.java
@@ -1,0 +1,10 @@
+package fuzs.bagofholding.api.config;
+
+public interface ContainerItemTooltipConfig {
+
+    boolean colorfulTooltips();
+
+    boolean contentsRequireShift();
+
+    boolean slotOverlay();
+}