+++ Common/src/main/java/fuzs/bagofholding/api/world/inventory/SimpleContainerWithSlots.java
@@ -1,4 +1,4 @@
-package fuzs.bagofholding.world.inventory;
+package fuzs.bagofholding.api.world.inventory;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
@@ -9,16 +9,17 @@ import net.minecraft.world.item.ItemStack;
  * copied from ender chest, as over there only difference from simple container is that slot ids are saved along with items
  */
 public class SimpleContainerWithSlots extends SimpleContainer {
+
     public SimpleContainerWithSlots(int containerRows) {
         super(containerRows * 9);
     }
 
     @Override
     public void fromTag(ListTag p_40108_) {
-        for (int i = 0; i < this.getContainerSize(); ++i) {
+        for(int i = 0; i < this.getContainerSize(); ++i) {
             this.setItem(i, ItemStack.EMPTY);
         }
-        for (int k = 0; k < p_40108_.size(); ++k) {
+        for(int k = 0; k < p_40108_.size(); ++k) {
             CompoundTag compoundtag = p_40108_.getCompound(k);
             int j = compoundtag.getByte("Slot") & 255;
             if (j >= 0 && j < this.getContainerSize()) {
@@ -32,7 +33,7 @@ public class SimpleContainerWithSlots extends SimpleContainer {
     public ListTag createTag() {
         ListTag listtag = new ListTag();
 
-        for (int i = 0; i < this.getContainerSize(); ++i) {
+        for(int i = 0; i < this.getContainerSize(); ++i) {
             ItemStack itemstack = this.getItem(i);
             if (!itemstack.isEmpty()) {
                 CompoundTag compoundtag = new CompoundTag();