+++ Common/src/main/java/fuzs/bagofholding/mixin/ServerPlaceRecipeMixin.java
@@ -7,6 +7,7 @@ import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
@@ -16,14 +17,15 @@ public abstract class ServerPlaceRecipeMixin {
     protected Inventory inventory;
 
     @ModifyVariable(method = "moveItemToGrid", at = @At("STORE"), ordinal = 0)
-    protected int moveItemToGrid$storeIndex(int itemIndex, Slot p_135439_, ItemStack p_135440_) {
+    protected int bagofholding$moveItemToGrid(int itemIndex, Slot p_135439_, ItemStack p_135440_) {
         if (p_135440_.getItem() instanceof RecipesIgnoreTag) {
-            return this.findSlotMatchingItem(this.inventory, p_135440_);
+            return this.bagofholding$findSlotMatchingItem(this.inventory, p_135440_);
         }
         return itemIndex;
     }
 
-    private int findSlotMatchingItem(Inventory inventory, ItemStack p_36031_) {
+    @Unique
+    private int bagofholding$findSlotMatchingItem(Inventory inventory, ItemStack p_36031_) {
         for(int i = 0; i < inventory.items.size(); ++i) {
             // vanilla uses many more checks here for tag/damage/enchantment/name
             if (!inventory.items.get(i).isEmpty() && p_36031_.is(inventory.items.get(i).getItem())) {