+++ Common/src/main/java/fuzs/bagofholding/mixin/client/TutorialMixin.java
@@ -1,6 +1,6 @@
 package fuzs.bagofholding.mixin.client;
 
-import fuzs.bagofholding.client.tutorial.ContainerItemTutorial;
+import fuzs.bagofholding.api.client.tutorial.ContainerItemTutorial;
 import fuzs.bagofholding.world.item.BagOfHoldingItem;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.Options;
@@ -20,12 +20,12 @@ public abstract class TutorialMixin {
     private ContainerItemTutorial bagItemTutorial;
 
     @Inject(method = "<init>", at = @At("TAIL"))
-    public void init(Minecraft p_175022_, Options p_175023_, CallbackInfo callbackInfo) {
+    public void bagofholding$init(Minecraft p_175022_, Options p_175023_, CallbackInfo callback) {
         this.bagItemTutorial = new ContainerItemTutorial((Tutorial) (Object) this, p_175023_, stack -> stack.getItem() instanceof BagOfHoldingItem, Component.translatable("tutorial.container.bag_of_holding.name"));
     }
 
     @Inject(method = "onInventoryAction", at = @At("TAIL"))
-    public void onInventoryAction(ItemStack p_175025_, ItemStack p_175026_, ClickAction p_175027_, CallbackInfo callbackInfo) {
+    public void bagofholding$onInventoryAction(ItemStack p_175025_, ItemStack p_175026_, ClickAction p_175027_, CallbackInfo callback) {
         this.bagItemTutorial.onInventoryAction(p_175025_, p_175026_, p_175027_);
     }
 }